pyproject.toml
[project]
name = "duld"
version = "0.1.0"
description = "Drive UpLoad Daemon"
license = "MIT"
readme = "README.md"
authors = [
    {name = "Wei-Cheng Pan", email = "legnaleurc@gmail.com"}
]
dynamic = ["dependencies", "requires-python"]

[tool.poetry]
requires-poetry = "^2.1"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0.1"
aiohttp = "^3.12.14"
wcpan-drive-cli = "^3.7.3"
wcpan-drive-core = "^4.2.0"
asyncinotify = "^4.2.1"
wcpan-drive-google = "^9.1.7"
wcpan-drive-crypt = "^4.1.2"
transmission-rpc = "^4.1.2"
wcpan-logging = "^2.0.2"
dacite = "^1.8.0"
wcpan-drive-sqlite = "^2.0.0"
qbittorrent-api = "^2025.7.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
